fix(rename): all plugins in the frontend renamed into extension
