+++ aqueductcore/frontend/src/components/atoms/IntegerField/index.tsx
@@ -1,6 +1,6 @@
 import { Box, Grid, TextField, TextFieldProps, styled } from "@mui/material";
 
-import { FieldDescription, FieldTitle, FieldType } from "components/atoms/sharedStyledComponents/pluginInputFields"
+import { FieldDescription, FieldTitle, FieldType } from "components/atoms/sharedStyledComponents/ExtensionInputFields"
 import { ExtensionFieldBase } from "types/globalTypes";
 
 const IntegerInput = styled(TextField)`
