+++ aqueductcore/frontend/src/components/atoms/TextAreaField/index.tsx
@@ -1,6 +1,6 @@
 import { Box, Grid, TextareaAutosize, TextareaAutosizeProps, styled } from "@mui/material";
 
-import { FieldDescription, FieldTitle, FieldType } from "components/atoms/sharedStyledComponents/pluginInputFields"
+import { FieldDescription, FieldTitle, FieldType } from "components/atoms/sharedStyledComponents/ExtensionInputFields"
 import { ExtensionFieldBase } from "types/globalTypes";
 
 const TextAreaInput = styled(TextareaAutosize)`
